aún no resueltas las restricciones

+++ directorio/views.py
@@ -6,12 +6,14 @@ from django.contrib.auth import login, logout
 from django.contrib.auth.forms import AuthenticationForm
 from django.db.models import Q
 from django.contrib.auth.decorators import login_required, user_passes_test
+from .decorators import custom_login_required
 
 def es_superusuario(user):
     return user.is_superuser
 
 @login_required
 def lista_registros(request):
+
     user = request.user
     registros = Registro.objects.all()
 
@@ -25,8 +27,12 @@ def lista_registros(request):
 
     return render(request, 'directorio/lista_registros.html', {'registros': registros, 'user': request.user})
 
-@user_passes_test(es_superusuario)
+@custom_login_required
 def agregar_registro(request):
+
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Inicie sesión como administrador para acceder a esta ventana.")
+        return redirect('login')
 
     if request.method == 'POST':
         form = RegistroForm(request.POST)
@@ -40,6 +46,11 @@ def agregar_registro(request):
 
 @user_passes_test(es_superusuario)
 def editar_registro(request, numero_registro):
+
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Inicie sesión como administrador para acceder a esta ventana.")
+        return redirect('login')
+
     registro = get_object_or_404(Registro, numero_registro=numero_registro)
 
     if request.method == 'POST':
@@ -54,6 +65,11 @@ def editar_registro(request, numero_registro):
 
 @user_passes_test(es_superusuario)
 def eliminar_registro(request, numero_registro):
+
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Inicie sesión como administrador para acceder a esta ventana.")
+        return redirect('login')
+
     registro = get_object_or_404(Registro, numero_registro=numero_registro)
     if request.method == 'POST':
         registro.delete()
@@ -72,21 +88,15 @@ def registroUsuario_view(request):
     
     return render(request, 'registration/registroUsuario.html', {'form': form})
 
+@login_required
 def login_view(request):
+
     if request.method == 'POST':
         form = AuthenticationForm(request, data=request.POST)
         if form.is_valid():
             user = form.get_user()
-            login(request, user)
-
-            next_url = request.GET.get('?next=')
-            if next_url and not user.is_superuser:
-                # Mostrar un mensaje de error
-                messages.error(request, "Acceso restringido. Esta función es solo para superusuarios.")
-                # Redirigir a una página específica en lugar de volver al formulario de inicio de sesión
-                return redirect('directorio/lista_registros')
-            
-            return redirect('directorio/lista_registros')  # Redirigir a la página de lista_registros después del inicio de sesión
+            login(request, user)           
+            return redirect('directorio:lista_registros') # Redirigir a la página de lista_registros después del inicio de sesión
     else:
         form = AuthenticationForm()
     return render(request, 'registration/login.html', {'form':form})
